BUGFIX -> if the user is not loggedIn redirect/browse to login page and vice versa
