+++ README.md
@@ -15,6 +15,11 @@
     - Created redux Store with userSlice
         - used onAuthStateChanged api
     - learnt aboout useNavigate() hook
+    - Implemented Signout
+    - UpdateProfileApi call also
+    - Fetch Movies from TMBD movies
+    - BugFix -> if the user is loggedIn, redirect him to the browse page
+    - BUGFIX -> if the user is not loggedIn redirect/browse to login page and vice versa
 # Features
 
 - Login/SignUp
